move all output-related code into it's own file
